backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
click==8.0.1
fastapi==0.65.1
h11==0.12.0
loguru==0.5.3
mongoengine==0.23.1
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
regex==2021.7.6
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
yapf==0.31.0
